+++ stac_fastapi/api/stac_fastapi/api/app.py
@@ -1,4 +1,4 @@
-"""fastapi app creation."""
+"""Fastapi app creation."""
 from typing import Any, Dict, List, Optional, Tuple, Type, Union
 
 import attr
@@ -38,22 +38,32 @@ from stac_fastapi.types.search import BaseSearchGetRequest, BaseSearchPostReques
 class StacApi:
     """StacApi factory.
 
-    Factory for creating a STAC-compliant FastAPI application.  After instantation, the application is accessible from
-    the `StacApi.app` attribute.
+    Factory for creating a STAC-compliant FastAPI application.  After
+    instantation, the application is accessible from the `StacApi.app` attribute.
 
     Attributes:
         settings:
-            API settings and configuration, potentially using environment variables. See https://pydantic-docs.helpmanual.io/usage/settings/.
+            API settings and configuration, potentially using environment
+            variables. See https://pydantic-docs.helpmanual.io/usage/settings/.
         client:
-            A subclass of `stac_api.clients.BaseCoreClient`.  Defines the application logic which is injected into the API.
+            A subclass of `stac_api.clients.BaseCoreClient`.  Defines the
+            application logic which is injected into the API.
         extensions:
-            API extensions to include with the application.  This may include official STAC extensions as well as third-party add ons.
+            API extensions to include with the application.  This may include
+            official STAC extensions as well as third-party add ons.
         exceptions:
-            Defines a global mapping between exceptions and status codes, allowing configuration of response behavior on certain exceptions (https://fastapi.tiangolo.com/tutorial/handling-errors/#install-custom-exception-handlers).
+            Defines a global mapping between exceptions and status codes,
+            allowing configuration of response behavior on certain exceptions
+            (https://fastapi.tiangolo.com/tutorial/handling-errors/#install-custom-exception-handlers).
         app:
             The FastAPI application, defaults to a fresh application.
-        route_dependencies (list of tuples of route scope dicts (eg `{'path': '/collections', 'method': 'POST'}`) and list of dependencies (e.g. `[Depends(oauth2_scheme)]`)):
-            Applies specified dependencies to specified routes. This is useful for applying custom auth requirements to routes defined elsewhere in the application.
+        route_dependencies:
+            List of tuples of route scope dicts (eg `{'path':
+            '/collections', 'method': 'POST'}`) and list of dependencies (e.g.
+            `[Depends(oauth2_scheme)]`)).  Applies specified dependencies to
+            specified routes. This is useful
+            for applying custom auth requirements to routes defined elsewhere in
+            the application.
     """
 
     settings: ApiSettings = attr.ib()
@@ -346,8 +356,11 @@ class StacApi:
         """Add custom dependencies to routes.
 
         Args:
-            scopes: list of scopes. Each scope should be a dict with a `path` and `method` property.
-            dependencies: list of [FastAPI dependencies](https://fastapi.tiangolo.com/tutorial/dependencies/) to apply to each scope.
+            scopes: list of scopes. Each scope should be a dict with a `path`
+                and `method` property.
+            dependencies: list of [FastAPI
+                dependencies](https://fastapi.tiangolo.com/tutorial/dependencies/)
+                to apply to each scope.
 
         Returns:
             None