+++ stac_fastapi/types/stac_fastapi/types/__init__.py
@@ -1,1 +1,1 @@
-"""backend submodule."""
+"""Backend submodule."""