+++ stac_fastapi/types/stac_fastapi/types/conformance.py
@@ -5,10 +5,10 @@ from enum import Enum
 class STACConformanceClasses(str, Enum):
     """Conformance classes for the STAC API spec."""
 
-    CORE = "https://api.stacspec.org/v1.0.0-rc.1/core"
-    OGC_API_FEAT = "https://api.stacspec.org/v1.0.0-rc.1/ogcapi-features"
-    COLLECTIONS = "https://api.stacspec.org/v1.0.0-rc.1/collections"
-    ITEM_SEARCH = "https://api.stacspec.org/v1.0.0-rc.1/item-search"
+    CORE = "https://api.stacspec.org/v1.0.0/core"
+    OGC_API_FEAT = "https://api.stacspec.org/v1.0.0/ogcapi-features"
+    COLLECTIONS = "https://api.stacspec.org/v1.0.0/collections"
+    ITEM_SEARCH = "https://api.stacspec.org/v1.0.0/item-search"
 
 
 class OAFConformanceClasses(str, Enum):