+++ stac_fastapi/types/stac_fastapi/types/core.py
@@ -28,19 +28,21 @@ class BaseTransactionsClient(abc.ABC):
 
     @abc.abstractmethod
     def create_item(
-        self, collection_id: str, item: stac_types.Item, **kwargs
-    ) -> Optional[Union[stac_types.Item, Response]]:
+        self,
+        collection_id: str,
+        item: Union[stac_types.Item, stac_types.ItemCollection],
+        **kwargs,
+    ) -> Optional[Union[stac_types.Item, Response, None]]:
         """Create a new item.
 
         Called with `POST /collections/{collection_id}/items`.
 
         Args:
-            item: the item
+            item: the item or item collection
             collection_id: the id of the collection from the resource path
 
         Returns:
-            The item that was created.
-
+            The item that was created or None if item collection.
         """
         ...
 
@@ -50,9 +52,10 @@ class BaseTransactionsClient(abc.ABC):
     ) -> Optional[Union[stac_types.Item, Response]]:
         """Perform a complete update on an existing item.
 
-        Called with `PUT /collections/{collection_id}/items`. It is expected that this item already exists.  The update
-        should do a diff against the saved item and perform any necessary updates.  Partial updates are not supported
-        by the transactions extension.
+        Called with `PUT /collections/{collection_id}/items`. It is expected
+        that this item already exists.  The update should do a diff against the
+        saved item and perform any necessary updates.  Partial updates are not
+        supported by the transactions extension.
 
         Args:
             item: the item (must be complete)
@@ -102,8 +105,9 @@ class BaseTransactionsClient(abc.ABC):
     ) -> Optional[Union[stac_types.Collection, Response]]:
         """Perform a complete update on an existing collection.
 
-        Called with `PUT /collections`. It is expected that this item already exists.  The update should do a diff
-        against the saved collection and perform any necessary updates.  Partial updates are not supported by the
+        Called with `PUT /collections`. It is expected that this item already
+        exists.  The update should do a diff against the saved collection and
+        perform any necessary updates.  Partial updates are not supported by the
         transactions extension.
 
         Args:
@@ -138,18 +142,21 @@ class AsyncBaseTransactionsClient(abc.ABC):
 
     @abc.abstractmethod
     async def create_item(
-        self, collection_id: str, item: stac_types.Item, **kwargs
-    ) -> Optional[Union[stac_types.Item, Response]]:
+        self,
+        collection_id: str,
+        item: Union[stac_types.Item, stac_types.ItemCollection],
+        **kwargs,
+    ) -> Optional[Union[stac_types.Item, Response, None]]:
         """Create a new item.
 
         Called with `POST /collections/{collection_id}/items`.
 
         Args:
-            item: the item
+            item: the item or item collection
+            collection_id: the id of the collection from the resource path
 
         Returns:
-            The item that was created.
-
+            The item that was created or None if item collection.
         """
         ...
 
@@ -159,9 +166,10 @@ class AsyncBaseTransactionsClient(abc.ABC):
     ) -> Optional[Union[stac_types.Item, Response]]:
         """Perform a complete update on an existing item.
 
-        Called with `PUT /collections/{collection_id}/items`. It is expected that this item already exists.  The update
-        should do a diff against the saved item and perform any necessary updates.  Partial updates are not supported
-        by the transactions extension.
+        Called with `PUT /collections/{collection_id}/items`. It is expected
+        that this item already exists.  The update should do a diff against the
+        saved item and perform any necessary updates.  Partial updates are not
+        supported by the transactions extension.
 
         Args:
             item: the item (must be complete)
@@ -210,8 +218,9 @@ class AsyncBaseTransactionsClient(abc.ABC):
     ) -> Optional[Union[stac_types.Collection, Response]]:
         """Perform a complete update on an existing collection.
 
-        Called with `PUT /collections`. It is expected that this item already exists.  The update should do a diff
-        against the saved collection and perform any necessary updates.  Partial updates are not supported by the
+        Called with `PUT /collections`. It is expected that this item already
+        exists.  The update should do a diff against the saved collection and
+        perform any necessary updates.  Partial updates are not supported by the
         transactions extension.
 
         Args:
@@ -280,7 +289,7 @@ class LandingPageMixin(abc.ABC):
                 {
                     "rel": Relations.conformance.value,
                     "type": MimeTypes.json,
-                    "title": "STAC/WFS3 conformance classes implemented by this server",
+                    "title": "STAC/OGC conformance classes implemented by this server",
                     "href": urljoin(base_url, "conformance"),
                 },
                 {
@@ -318,7 +327,8 @@ class BaseCoreClient(LandingPageMixin, abc.ABC):
     post_request_model = attr.ib(default=BaseSearchPostRequest)
 
     def conformance_classes(self) -> List[str]:
-        """Generate conformance classes by adding extension conformance to base conformance classes."""
+        """Generate conformance classes by adding extension conformance to base
+        conformance classes."""
         base_conformance_classes = self.base_conformance_classes.copy()
 
         for extension in self.extensions:
@@ -525,7 +535,8 @@ class AsyncBaseCoreClient(LandingPageMixin, abc.ABC):
     post_request_model = attr.ib(default=BaseSearchPostRequest)
 
     def conformance_classes(self) -> List[str]:
-        """Generate conformance classes by adding extension conformance to base conformance classes."""
+        """Generate conformance classes by adding extension conformance to base
+        conformance classes."""
         conformance_classes = self.base_conformance_classes.copy()
 
         for extension in self.extensions:
@@ -718,12 +729,11 @@ class AsyncBaseFiltersClient(abc.ABC):
     ) -> Dict[str, Any]:
         """Get the queryables available for the given collection_id.
 
-        If collection_id is None, returns the intersection of all
-        queryables over all collections.
+        If collection_id is None, returns the intersection of all queryables over all
+        collections.
 
         This base implementation returns a blank queryable schema. This is not allowed
         under OGC CQL but it is allowed by the STAC API Filter Extension
-
         https://github.com/radiantearth/stac-api-spec/tree/master/fragments/filter#queryables
         """
         return {
@@ -745,13 +755,12 @@ class BaseFiltersClient(abc.ABC):
     ) -> Dict[str, Any]:
         """Get the queryables available for the given collection_id.
 
-        If collection_id is None, returns the intersection of all
-        queryables over all collections.
+        If collection_id is None, returns the intersection of all queryables over all
+        collections.
 
         This base implementation returns a blank queryable schema. This is not allowed
         under OGC CQL but it is allowed by the STAC API Filter Extension
-
-        https://github.com/radiantearth/stac-api-spec/tree/master/fragments/filter#queryables
+        https://github.com/stac-api-extensions/filter#queryables
         """
         return {
             "$schema": "https://json-schema.org/draft/2019-09/schema",